Added Auth Protection to all user and post routes

+++ app/api/v1/posts.py
@@ -5,6 +5,8 @@ from schemas.post_schema import PostCreate, PostResponse, PostUpdate
 from sqlalchemy import func, select
 from sqlalchemy.orm import Session
 
+from app.utils.auth import CurrentUser
+
 from db.database import get_db
 
 post_router = APIRouter(prefix="/api/v1/posts", tags=["posts"])
@@ -24,18 +26,15 @@ def get_posts(db: Annotated[Session, Depends(get_db)]):
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-def create_post(post: PostCreate, db: Annotated[Session, Depends(get_db)]):
-    result = db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+def create_post(
+    post: PostCreate,
+    current_user: CurrentUser, 
+    db: Annotated[Session, Depends(get_db)]
+):
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     db.commit()
@@ -56,6 +55,7 @@ def get_post(post_id: int, db: Annotated[Session, Depends(get_db)]):
 def update_post_full(
     post_id: int, 
     post_data: PostCreate, 
+    current_user: CurrentUser,
     db: Annotated[Session, Depends(get_db)]
 ):
     result = db.execute(select(models.Post).where(models.Post.id == post_id))
@@ -63,20 +63,14 @@ def update_post_full(
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
-    if post_data.user_id != post.user_id:
-        result = db.execute(
-            select(models.User).where(models.User.id == post_data.user_id)
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not Authorized to update this post",
         )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
         
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     db.commit()
     db.refresh(post)
@@ -84,15 +78,22 @@ def update_post_full(
 
 
 @post_router.patch("/{post_id}", response_model=PostResponse)
-def update_post_full(
+def update_post_partial(
     post_id: int, 
     post_data: PostUpdate, 
+    current_user: CurrentUser,
     db: Annotated[Session, Depends(get_db)]
 ):
     result = db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not Authorized to update this post"
+        )
 
     update_data = post_data.model_dump(exclude_unset=True)
     for field, value in update_data.items():
@@ -104,12 +105,21 @@ def update_post_full(
 
 
 @post_router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-def get_post(post_id: int, db: Annotated[Session, Depends(get_db)]):
+def get_post(
+    post_id: int, 
+    current_user: CurrentUser,
+    db: Annotated[Session, Depends(get_db)]
+):
     result = db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not Authorized to delete this post"
+        )
 
     db.delete(post)
     db.commit()
